requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
bcrypt
httpx
google-auth-oauthlib
google-api-python-client
msal
caldav
cryptography
python-dotenv
jinja2==3.1.4
python-multipart
itsdangerous
pydantic-settings
psycopg2-binary
requests
